fix: Address the review's three correctness findings
Widens both on-disk digests from 32 to 64 bits and folds the source's
byte length into the source digest. Each is the sole guard on a
correctness decision -- whether a payload still describes the source,
and whether a record survived the trip to disk -- and at 32 bits a
birthday collision is within reach of an ordinary project's worth of
sources, which serves a stale payload as a hit. The record layout
changed, so CACHE_FORMAT_VERSION goes to 2 and the fixtures are
regenerated.

Checks for a duplicate key before consulting the record's version tag,
so a duplicate whose second record is version-stale is still reported as
the writer bug it is rather than quietly deduplicated. Adds
duplicate_key_across_versions_store.bin and its test.

Creates the BSCache singleton through a function-local static and routes
every accessor through get_singleton(). Godot loads scripts from worker
threads, and the previous null check could construct two caches and
leave two callers locking different mutexes over the same maps.

+++ src/barista_script_parse_cache.cpp
@@ -146,8 +146,8 @@ godot::String BaristaScriptParseCache::get_default_store_path() {
 	return BSParseCache::get_default_store_path();
 }
 
-int BaristaScriptParseCache::compute_source_digest(const godot::String &p_source) {
-	return (int)BSParseCache::compute_source_digest(p_source);
+int64_t BaristaScriptParseCache::compute_source_digest(const godot::String &p_source) {
+	return (int64_t)BSParseCache::compute_source_digest(p_source);
 }
 
 godot::String BaristaScriptParseCache::get_source_code(const godot::String &p_path) {
